+++ tests/unit/test_ssh.py
@@ -1,6 +1,6 @@
 import pytest
 
-from vault_cli import ssh
+from vault_cli import exceptions, ssh
 
 
 @pytest.fixture
@@ -24,8 +24,20 @@ def test_ensure_agent_already_present(exec, environ):
 # This should rather be an integration test
 def test_launch_command(environ):
     environ.update({"A": "B"})
-    stdout = ssh._launch_command(["env"], "", {"C": "D"})
+    stdout = ssh._launch_command(command=["env"], stdin="", environment={"C": "D"})
     assert "A=B\nC=D" in stdout
+
+
+# This should rather be an integration test
+def test_launch_command_error(environ):
+    non_existing_filename = "a" * 30
+    with pytest.raises(exceptions.VaultSubprocessException) as exc_info:
+        ssh._launch_command(
+            command=["cat", non_existing_filename], stdin="", environment={}
+        )
+
+    error_message = f"cat: {non_existing_filename}: No such file or directory"
+    assert str(exc_info.value) == error_message
 
 
 @pytest.fixture